Add tests for global configuration and support mail tests
