+++ package.json
@@ -105,6 +105,7 @@
     "rollup-plugin-vue": "^6.0.0",
     "rtlcss": "^3.5.0",
     "sass-embedded": "^1.59.3",
+    "smtp-tester": "^2.0.1",
     "stylelint": "^14.16.1",
     "stylelint-config-standard": "^24.0.0",
     "stylelint-order": "^5.0.0",
