+++ tests/System/support/index.js
@@ -2,6 +2,7 @@ import './commands';
 import 'joomla-cypress';
 
 before(() => {
+  cy.task('startMailServer');
   Cypress.on('uncaught:exception', (err, runnable) => {
     console.log(`err :${err}`);
     console.log(`runnable :${runnable}`);
